refactor: move card data service to separate file
