+++ YouTube-client-app/src/app/core/services/search-service/search.service.ts
@@ -7,7 +7,7 @@ import {
 import { getCardsSuccess } from '../../../redux/actions/card.actions';
 import { selectCombinedCards } from '../../../redux/selectors/card.selector';
 import { CardModel } from '../../../redux/state.model';
-import { CardDataService } from '../../../youtube/services/card-data.service';
+import { CardDataService } from '../../../youtube/services/card-data-service/card-data.service';
 
 @Injectable({
   providedIn: 'root',
