+++ YouTube-client-app/src/app/redux/effects/card.effects.ts
@@ -3,7 +3,7 @@ import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { of } from 'rxjs';
 import { catchError, map, switchMap } from 'rxjs/operators';
 
-import { CardDataService } from '../../youtube/services/card-data.service';
+import { CardDataService } from '../../youtube/services/card-data-service/card-data.service';
 import {
   getCards,
   getCardsFailed,
