+++ YouTube-client-app/src/app/youtube/services/card-data-service/card-data.service.fixtures.ts
@@ -1,4 +1,4 @@
-import { CardsListModel } from '../models/cards-list.model';
+import { CardsListModel } from '../../models/cards-list.model';
 
 export const mockCardsDataResponse: CardsListModel = {
   kind: 'testKind',
